Gravity implemented. Double jump available.

+++ Gio/core/src/com/ihcreations/gio/characters/CharacterBody.java
@@ -1,7 +1,7 @@
 package com.ihcreations.gio.characters;
 
 // Java Imports
-import java.awt.Point;
+import com.ihcreations.gio.utils.Point;
 
 // Libgdx Imports
 import com.badlogic.gdx.graphics.Texture;
@@ -9,8 +9,8 @@ import com.badlogic.gdx.graphics.Texture;
 public class CharacterBody {
 	// Public Member Variables
 	public Point m_center;
-	public int m_height;
-	public int m_width;
+	public float m_height;
+	public float m_width;
 	
 	// Private Member Variables
 	private Texture m_sprite;
@@ -22,14 +22,14 @@ public class CharacterBody {
 		this.m_sprite = null;
 	}
 	
-	public CharacterBody(Point center, int height, int width, Texture sprite) {
+	public CharacterBody(Point center, float height, float width, Texture sprite) {
 		this.m_height = height;
 		this.m_width = width;
 		this.m_center = center;
 		this.m_sprite = sprite;
 	}
 	
-	public CharacterBody(int x, int y, int height, int width, Texture sprite) {
+	public CharacterBody(float x, float y, float height, float width, Texture sprite) {
 		this.m_height = height;
 		this.m_width = width;
 		this.m_center = new Point(x, y);
